Fix bug: Validate platform before deciding Specs
